dreamlog-backend/requirements.txt
# ===== 核心框架 =====
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# ===== 数据库 =====
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pgvector==0.3.6

# ===== Redis =====
redis[hiredis]==5.2.1

# ===== 异步任务 =====
celery[redis]==5.4.0

# ===== 认证 =====
PyJWT==2.10.1
passlib[bcrypt]==1.7.4

# ===== AI 服务 =====
anthropic==0.43.0
openai==1.58.1

# ===== HTTP 客户端 =====
httpx==0.28.1

# ===== 图片处理 =====
Pillow==11.1.0

# ===== 对象存储 =====
oss2==2.19.1
boto3==1.36.2

# ===== 工具 =====
python-dotenv==1.0.1
python-multipart==0.0.20

# ===== 开发工具 =====
pytest==8.3.4
pytest-asyncio==0.25.0
black==24.10.0
ruff==0.8.6
